Knowledge and Skill contribution fails if the initial taxonomy clone is not present in user github account.
UI attempts to create the clone, but the clone can
take some time to finish. When UI attempts to get the
base sha of the repo's main branch it fails because
the repository is empty.

Add retry logic to check the bash sha every 5 seconds
and make 10 attempts. If clone takes more than 50 seconds
it fails.

In this specific scenario, when user creates it's first
contribution, it can take a bit longer because of the
repo cloning. Without any notification, this wait time
might look like a no-op to the user. To address this
added a spinner with an alter message that notifies
user that the submission is in progress.

Patch also includes refactoring of duplicate code present
across api/skill and api/knowledge.

Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>