Fine-tuning and post train chat eval POC
- Demonstrates an end to end knowledge submission, generate,
train and post-train side-by-side model comparison for the user
to validate their knowledge submission is included in the newly
trained checkpoint.
- For this to function for the frontend to make REST API calls to
Instructlab this uses an api-server that frontends ilab. The code
is here https://github.com/nerdalert/ilab-api-server
- The demo was run on a 24GB GPU leveraging the simple pipeline.
Will get an example acceslerated pipeline demo with some hardware
soon.
- Training and generation for the demo took around ~30-45m or so.
- All functionality is decoupled from the system via REST
making it serviceable out of the gate and enabling the UI
functionality.

Signed-off-by: Brent Salisbury <bsalisbu@redhat.com>