Fixes DCO signoff issue
Auth session keeps information from the user profile,
so if the email address is not made visible on the profile
or alias is set as an email address, UI was using that email
address for signoff. In some cases, this was failing because
that email address is not users primary email address.

For DCO to pass, user needs to signoff the commits with it's
primary email address set in his account.

Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>