fix: Mergify and Dependabot configs YAML are incorrect/inconsistent
Some of the YAML configurations are inconsistent and/or incorrect (and are not being actively parsed).

To ensure these files continue to be parsed, I have also introduced a YAML linter for the `.github` directory. There might be a bit of work to lint other YAML files. That is out of scope for this PR, though.

Signed-off-by: Courtney Pacheco <6019922+courtneypacheco@users.noreply.github.com>